Add client config library
