--- pkg/v1/config/doc.go
@@ -1,5 +1,0 @@
-// Copyright 2021 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Package config provides functions for reading and writing client configuration.
-package config
